Urgent bugfix: Refactor CRC variables from unsigned long to unsigned int to make it work properly on 64-bit machines (with 64-bit Python and libraries). Fixes #1146.

+++ source/include/crc.h
@@ -21,7 +21,7 @@
 #ifndef __CRC_H
 #define __CRC_H
 
-extern unsigned long crc_table[256];
+extern unsigned int crc_table[256];
 
 #define INIT_CRC(x) (x) = 0xffffffff
 #define FINISH_CRC(x) (x) = (x) ^ 0xffffffff
@@ -31,19 +31,19 @@ extern unsigned long crc_table[256];
       b != e; \
       crc = (crc >> 8) ^ crc_table[(crc & 0xFF) ^ *(b++)]); \
 
-inline void add_CRC(const int data, unsigned long &crc)
+inline void add_CRC(const int data, unsigned int &crc)
 { ADD_CRC }
 
-inline void add_CRC(const unsigned long data, unsigned long &crc)
+inline void add_CRC(const unsigned long data, unsigned int &crc)
 { ADD_CRC }
 
-inline void add_CRC(const float data, unsigned long &crc)
+inline void add_CRC(const float data, unsigned int &crc)
 { ADD_CRC }
 
-inline void add_CRC(const unsigned char c, unsigned long &crc)
+inline void add_CRC(const unsigned char c, unsigned int &crc)
 { crc = (crc >> 8) ^ crc_table[(crc & 0xFF) ^ c]; }
 
-inline void add_CRC(const char *c, unsigned long &crc)
+inline void add_CRC(const char *c, unsigned int &crc)
 {
   for(; *c; add_CRC((unsigned char)*c++, crc));
   add_CRC((unsigned char)0, crc);
